Messi i ToInt e sistemato OrdineValidator

+++ Codice/Backend/Validators/filialeValidator.ts
@@ -17,19 +17,22 @@ function indirizzoValidator(chain: ValidationChain): ValidationChain {
 function num_tavoliValidator(chain: ValidationChain): ValidationChain {
   return chain
         .notEmpty().withMessage('Il numero di tavoli è obbligatorio!')
-        .isInt({ min: 1 }).withMessage('Il numero di tavoli deve essere un numero intero positivo');
+        .isInt({ min: 1 }).withMessage('Il numero di tavoli deve essere un numero intero positivo')
+        .toInt();
 }
 
 function longitudineValidator(chain: ValidationChain): ValidationChain {
   return chain
         .notEmpty().withMessage('La longitudine è obbligatoria!')
-        .isFloat({ min: -180, max: 180 }).withMessage('La longitudine deve essere un numero tra -180 e 180');
+        .isFloat({ min: -180, max: 180 }).withMessage('La longitudine deve essere un numero tra -180 e 180')
+        .toFloat();
 }
 
 function latitudineValidator(chain: ValidationChain): ValidationChain {
   return chain
         .notEmpty().withMessage('La latitudine è obbligatoria!')
-        .isFloat({ min: -90, max: 90 }).withMessage('La latitudine deve essere un numero tra -90 e 90');
+        .isFloat({ min: -90, max: 90 }).withMessage('La latitudine deve essere un numero tra -90 e 90')
+        .toFloat();
 }
 
 function immagineValidator(chain: ValidationChain): ValidationChain {
@@ -41,7 +44,8 @@ function immagineValidator(chain: ValidationChain): ValidationChain {
 function idFilialeValidator(chain: ValidationChain): ValidationChain {
   return chain
         .notEmpty().withMessage('L\'ID della filiale è obbligatorio!')
-        .isInt({ min: 1 }).withMessage('L\'ID della filiale deve essere un intero positivo!');
+        .isInt({ min: 1 }).withMessage('L\'ID della filiale deve essere un intero positivo!')
+        .toInt();
 }
 
 // Validatori
